First round of http://dartbug.com/15922: - Address warnings about 64-bit to 32-bit conversions. - Remove heap profiler.
R=asiva@google.com

Review URL: https://codereview.chromium.org//139043003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@31867 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/platform/utils.h
@@ -45,34 +45,34 @@ class Utils {
   }
 
   template<typename T>
-  static inline bool IsAligned(T x, int n) {
+  static inline bool IsAligned(T x, intptr_t n) {
     ASSERT(IsPowerOfTwo(n));
     return (x & (n - 1)) == 0;
   }
 
   template<typename T>
-  static inline bool IsAligned(T* x, int n) {
+  static inline bool IsAligned(T* x, intptr_t n) {
     return IsAligned(reinterpret_cast<uword>(x), n);
   }
 
   template<typename T>
-  static inline T RoundDown(T x, int n) {
+  static inline T RoundDown(T x, intptr_t n) {
     ASSERT(IsPowerOfTwo(n));
     return (x & -n);
   }
 
   template<typename T>
-  static inline T* RoundDown(T* x, int n) {
+  static inline T* RoundDown(T* x, intptr_t n) {
     return reinterpret_cast<T*>(RoundDown(reinterpret_cast<uword>(x), n));
   }
 
   template<typename T>
-  static inline T RoundUp(T x, int n) {
+  static inline T RoundUp(T x, intptr_t n) {
     return RoundDown(x + n - 1, n);
   }
 
   template<typename T>
-  static inline T* RoundUp(T* x, int n) {
+  static inline T* RoundUp(T* x, intptr_t n) {
     return reinterpret_cast<T*>(RoundUp(reinterpret_cast<uword>(x), n));
   }
